requirements.txt
asgiref==3.6.0
asttokens==2.2.1
attrs==22.2.0
autopep8==2.0.1
backcall==0.2.0
beautifulsoup4==4.11.1
certifi==2022.12.7
charset-normalizer==3.0.1
colorama==0.4.6
comm==0.1.2
contourpy==1.0.7
cycler==0.11.0
debugpy==1.6.6
decorator==5.1.1
Django==4.2
djangorestframework==3.14.0
djangorestframework-simplejwt==5.2.2
executing==1.2.0
fonttools==4.38.0
idna==3.4
iniconfig==2.0.0
ipykernel==6.20.2
ipython==8.9.0
jedi==0.18.2
jupyter_client==8.0.2
jupyter_core==5.2.0
kiwisolver==1.4.4
matplotlib==3.6.3
matplotlib-inline==0.1.6
nest-asyncio==1.5.6
numpy==1.24.1
packaging==23.0
pandas==1.5.3
parso==0.8.3
pickleshare==0.7.5
Pillow==9.4.0
platformdirs==2.6.2
pluggy==1.0.0
prompt-toolkit==3.0.36
psutil==5.9.4
pure-eval==0.2.2
pycodestyle==2.10.0
Pygments==2.14.0
PyJWT==2.6.0
pyparsing==3.0.9
pytest==7.2.2
pytest-django==4.5.2
pytest-testdox==3.0.1
python-dateutil==2.8.2
pytz==2023.3
pywin32==305
pyzmq==25.0.0
requests==2.28.2
seaborn==0.12.2
six==1.16.0
soupsieve==2.3.2.post1
sqlparse==0.4.3
stack-data==0.6.2
tornado==6.2
traitlets==5.9.0
tzdata==2023.3
urllib3==1.26.14
wcwidth==0.2.6
